LayoutTests/media/track/track-cue-rendering-with-padding-expected.txt

Rendering text track cue line with padding set.

EVENT(canplaythrough)
EVENT(seeked)

** Cue should be displayed entirely within the container. **
EXPECTED (cueDisplayBottomPoint <= cueContainerBottomPoint == 'true') OK
END OF TEST
